Correction d'un bug dans le calcul du second moment. La fonction de sélection de la racine marche!

+++ trunk/VPTree/src/vector.cpp
@@ -75,12 +75,13 @@ double Vector::compute_median()
 
 double Vector::compute_second_moment(double val)
 {
-	int res = 0;
+	double res = 0;
 	for (unsigned int i= 0; i<(*this).size(); ++i)
 	{
 		double temp = (*this)[i] - val;
 		res += temp*temp;
 	}
+
 	return res/this->size();
 }
